Removed component state variables for MemberDetails.js and implemented REDUX instead
Changed order of link on buttons for navBar so they redirect correctly
Created complete JSON object for signupDetails of new member registration to hold information of all textfields required for creating a new member.
Created a one for all reducer for the registration pages to update global redux state, it recieves the state name and the state value and updates the signupDetails Object accordingly.